carp: change cancel work function
Not too sure why at the moment but the cancelling of work was
using the cancel_delayed_work method which takes a different
struct to what is being used for the rest of the module. Use
cancel_work_sync instead for the time being and add a note as
a reminder to check it out later.